Refactor Files Structure
