+++ app/build.gradle
@@ -18,6 +18,13 @@ android {
     }
 }
 
+allprojects {
+    repositories {
+
+        maven { url 'https://jitpack.io' }
+    }
+}
+
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'com.android.support:appcompat-v7:28.0.0'
@@ -33,4 +40,6 @@ dependencies {
     implementation 'com.ablanco.zoomy:zoomy:1.1.0'
     implementation 'com.github.samanzamani.persiandate:PersianDate:0.7'
     implementation 'com.akexorcist:RoundCornerProgressBar:2.0.3'
+    implementation 'com.github.AnyChart:AnyChart-Android:1.1.2'
+    implementation 'com.jaredrummler:material-spinner:1.3.1'
 }
